requirements.txt
absl-py==1.0.0
affine==2.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==5.0.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cssselect2==0.5.0
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dm-tree==0.1.6
entrypoints==0.4
etils==1.1.1
executing==0.8.3
Fiona==1.8.21
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
geopandas==0.10.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphviz==0.19.1
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.16.0
imgaug==0.4.0
importlib-metadata==4.11.1
importlib-resources==5.4.0
imutils==0.5.4
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.6.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.11.2
pyparsing==3.0.7
pyproj==3.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
pyzmq==22.3.0
qtconsole==5.4.1
QtPy==2.3.1
rasterio==1.2.10
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Rtree==1.0.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
Shapely==1.8.1.post1
six==1.16.0
sklearn==0.0
snuggs==1.4.7
soupsieve==2.3.1
spatialite==0.0.3
stack-data==0.2.0
svgwrite==1.4.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.8.3
#tensorflow-examples===5bc9f1ed519146242db5e71f00d9d39d52a308c8-
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
tf-keras-vis==0.8.1
threadpoolctl==3.1.0
tifffile==2022.2.9
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing-extensions==4.1.1
-e git+https://github.com/jakeret/unet.git@f557a51b6f95aae6848cab6141e6cae573934bf8#egg=unet
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.7.0
